fix: 외부 도서 등록할 시 내용 일부 수정 가능하도록 변경

+++ src/main/java/com/nhnacademy/hello/dto/book/AladinBookDTO.java
@@ -12,7 +12,7 @@ public record AladinBookDTO(
         LocalDate pubDate,
         Long isbn13,
         String description,
-        int salesPoint,
+        Long salesPoint,
         String cover
 ) {
 }
